# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ROSA Regional Platform API
  description: Platform API for ROSA HCP regional cluster management
  version: 0.0.1
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v0
    description: API v0
paths:
...
  /resource_bundles:
    get:
      summary: List all resource bundles
      description: |
        Returns a paginated list of resource bundles.
        Supports filtering via search query and custom field selection.
      operationId: listResourceBundles
      tags:
        - ResourceBundles
      parameters:
        - name: page
          in: query
          description: Page number (1-indexed)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 0
            default: 100
        - name: search
          in: query
          description: SQL-like filter criteria on resource attributes
          schema:
            type: string
        - name: orderBy
          in: query
          description: SQL-like ordering (e.g., "name asc")
          schema:
            type: string
        - name: fields
          in: query
          description: Comma-separated list of fields to return
          schema:
            type: string
        - name: X-Operation-ID
          in: header
          description: Optional operation ID for tracking
          schema:
            type: string
      responses:
        '200':
          description: List of resource bundles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceBundleList'
        '401':
          description: Invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - user lacks required permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /accounts/{accountId}/enable:
    post:
      summary: Enable account for authorization
      description: |
        Enables an AWS account for Cedar/AVP-based authorization.
        Creates a policy store in Amazon Verified Permissions.
        Requires privileged access.
      operationId: enableAccount
      tags:
        - Authorization
      parameters:
        - name: accountId
          in: path
          required: true
          description: AWS account ID to enable
          schema:
            type: string
            pattern: '^\d{12}$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnableAccountRequest'
      responses:
        '201':
          description: Account enabled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - caller is not privileged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Account already enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /accounts/{accountId}/disable:
    post:
      summary: Disable account authorization
      description: |
        Disables an AWS account from authorization.
        Deletes the policy store and all associated policies, groups, and attachments.
        Requires privileged access.
      operationId: disableAccount
      tags:
        - Authorization
      parameters:
        - name: accountId
          in: path
          required: true
          description: AWS account ID to disable
          schema:
            type: string
            pattern: '^\d{12}$'
      responses:
        '204':
          description: Account disabled successfully
        '403':
          description: Forbidden - caller is not privileged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Authorization - Policy Management
...
components:
  schemas:
...
    ResourceBundle:
      type: object
      description: A resource bundle containing manifests for deployment
      required:
        - id
        - kind
        - href
        - name
      properties:
        id:
          type: string
          description: Unique identifier for the resource bundle
        kind:
          type: string
          description: Resource kind
          example: ResourceBundle
        href:
          type: string
          description: API href reference
        name:
          type: string
          description: Name of the resource bundle
        consumer_name:
          type: string
          description: Name of the consumer this bundle is associated with
        version:
          type: integer
          description: Version number of the resource bundle
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        deleted_at:
          type: string
          format: date-time
          description: Deletion timestamp (if soft-deleted)
        metadata:
          type: object
          description: Additional metadata
          additionalProperties:
            type: string
        manifests:
          type: array
          description: Array of Kubernetes manifests
          items:
            type: object
        delete_option:
          type: object
          description: Deletion strategy for the resource bundle
        manifest_configs:
          type: array
          description: Configuration for manifests
          items:
            type: object
        status:
          type: object
          description: Status of the resource bundle
    ResourceBundleList:
      type: object
      description: Paginated list of resource bundles
      required:
        - kind
        - page
        - size
        - total
        - items
      properties:
        kind:
          type: string
          description: Resource kind
          example: ResourceBundleList
        page:
          type: integer
          description: Current page number
        size:
          type: integer
          description: Number of items per page
        total:
          type: integer
          description: Total number of items
        items:
          type: array
          items:
            $ref: '#/components/schemas/ResourceBundle'
...
    Error:
      type: object
      description: Error response
      required:
        - code
        - reason
      properties:
        id:
          type: string
          description: Error ID
        kind:
          type: string
          description: Resource kind
          example: Error
        href:
          type: string
          description: API href reference
        code:
          type: string
          description: Error code
        reason:
          type: string
          description: Human-readable error message
        operation_id:
          type: string
          description: Request operation ID for tracing
...
    EnableAccountRequest:
      type: object
      description: Request body for enabling an account
      properties:
        privileged:
          type: boolean
          description: If true, account bypasses all authorization checks
          default: false
    Account:
      type: object
      description: An enabled account
      required:
        - accountId
        - privileged
        - createdAt
      properties:
        accountId:
          type: string
          description: AWS account ID
          pattern: '^\d{12}$'
        policyStoreId:
          type: string
          description: AVP policy store ID (null if privileged)
        privileged:
          type: boolean
          description: If true, bypasses all authorization
        createdAt:
          type: string
          format: date-time
          description: When the account was enabled
        createdBy:
          type: string
          description: ARN of who enabled the account
...