# pkg/handlers/resource_bundle.go
package handlers

import (
	"encoding/json"
	"log/slog"
	"net/http"
	"strconv"

	"github.com/openshift/rosa-regional-platform-api/pkg/clients/maestro"
	"github.com/openshift/rosa-regional-platform-api/pkg/middleware"
)

// ResourceBundleHandler handles resource bundle endpoints
type ResourceBundleHandler struct {
	maestroClient maestro.ClientInterface
	logger        *slog.Logger
}

// NewResourceBundleHandler creates a new ResourceBundleHandler
func NewResourceBundleHandler(maestroClient maestro.ClientInterface, logger *slog.Logger) *ResourceBundleHandler {
	return &ResourceBundleHandler{
		maestroClient: maestroClient,
		logger:        logger,
	}
}

// List handles GET /api/v0/resource_bundles
func (h *ResourceBundleHandler) List(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	accountID := middleware.GetAccountID(ctx)

	h.logger.Debug("listing resource bundles", "account_id", accountID)

	page := 1
	size := 100

	if p := r.URL.Query().Get("page"); p != "" {
		if parsed, err := strconv.Atoi(p); err == nil && parsed > 0 {
			page = parsed
		}
	}

	if s := r.URL.Query().Get("size"); s != "" {
		if parsed, err := strconv.Atoi(s); err == nil && parsed > 0 {
			size = parsed
		}
	}

	search := r.URL.Query().Get("search")
	orderBy := r.URL.Query().Get("orderBy")
	fields := r.URL.Query().Get("fields")

	list, err := h.maestroClient.ListResourceBundles(ctx, page, size, search, orderBy, fields)
	if err != nil {
		h.logger.Error("failed to list resource bundles from Maestro", "error", err, "account_id", accountID)
		if maestroErr, ok := err.(*maestro.Error); ok {
			h.writeError(w, http.StatusBadGateway, maestroErr.Code, maestroErr.Reason)
			return
		}
		h.writeError(w, http.StatusInternalServerError, "maestro-error", "Failed to list resource bundles")
		return
	}

	h.logger.Debug("resource bundles listed", "total", list.Total, "account_id", accountID)

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(list)
}

func (h *ResourceBundleHandler) writeError(w http.ResponseWriter, status int, code, reason string) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)

	resp := map[string]interface{}{
		"kind":   "Error",
		"code":   code,
		"reason": reason,
	}

	json.NewEncoder(w).Encode(resp)
}